requirements.txt
asgiref==3.7.2
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.0.0
Django==4.2.4
gunicorn==21.2.0
idna==3.4
opuslib==3.0.1
packaging==23.1
Pillow==10.0.0
pytesseract==0.3.10
requests==2.31.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uuid==1.30
whitenoise==6.5.0
